Rough UML upload
Here's a basic UML diagram, I had to abandon filling in everything
simply because of the amount of info I had to process.
I'll continue to orient myself with regards to the code and put in some basic validation checks,
which should be up by tomorrow at noon (at the latest).
Once that's done, I'll come back to the UML.

Shawn Ellis